chore: (#403) 파일 용도변경
